style: reorganization, get rid of pages component
