+++ src/components/ComposerChooser.js
