+++ src/components/PiqueText.js
